+++ app/api/auth/[...nextauth]/route.js
@@ -29,7 +29,10 @@ export const authOptions = {
                         password: credentials?.password,
                     });
 
-                    // console.log("=>>> after login successfully res", res);
+                    console.log(
+                        "=>>> after credentials login successfully res from api auth route page",
+                        res
+                    );
 
                     if (res.error) {
                         throw new Error("Email or Password is not correct");
@@ -59,14 +62,10 @@ export const authOptions = {
             },
         }),
     ],
-    pages: {
-        signIn: "/login",
-        signOut: "/login",
-    },
     secret: process.env.NEXTAUTH_SECRET,
     callbacks: {
         async signIn({ user, account, profile }) {
-            console.log("=>>> signIn...");
+            console.log("=>>> signIn from api auth route page ...");
 
             if (account.provider === "google") {
                 // Example accessToken and expiresIn, replace with actual token logic
@@ -77,8 +76,8 @@ export const authOptions = {
                 user.expiresIn = expiresIn;
                 user.email = profile?.email;
 
-                // console.log("user ----", user);
-                // console.log('profile ----', profile)
+                // console.log("user from api auth route page ----", user);
+                // console.log('profile from api auth route page ----', profile)
 
                 return { ...user };
             }
@@ -86,22 +85,20 @@ export const authOptions = {
         },
 
         async redirect({ url, baseUrl }) {
-            console.log("=>>> redirect...");
+            console.log("=>>> redirect from api auth route page ...");
 
             // const userExists = await checkUserExistByGoogleLogin({email: "robeul.starit@gmail.com"});
-            // console.log("userExists  1", userExists);
+            // console.log("userExists 1 from api auth route page", userExists);
 
             // if(userExists){
             //     if (userExists?.message == 'Already User Exists Account Provider Customer') {
-            //         console.log("userExists  2");
+            //         console.log("userExists 2 from api auth route page");
             //         return url.startsWith(baseUrl) ? `${baseUrl}/dashboard` : baseUrl;
             //     } else {
-            //         console.log("userExists  3");
+            //         console.log("userExists 3 from api auth route page");
             //         return url.startsWith(baseUrl) ? `${baseUrl}/google-profile` : baseUrl;
             //     }
             // }
-
-            
 
             // Allows relative callback URLs
             // if (url.startsWith("/")) return `${baseUrl}${url}`
@@ -109,14 +106,14 @@ export const authOptions = {
             // else if (new URL(url).origin === baseUrl) return url
             // return baseUrl
 
-            return url.startsWith(baseUrl) ? `${baseUrl}/dashboard` : baseUrl;
+            return url.startsWith(baseUrl) ? `${baseUrl}/new-user` : baseUrl;
         },
 
         async jwt({ token, user }) {
             console.log("=>>> jwt...");
 
-            // console.log('=>>> jwt token', token)
-            // console.log('=>>> jwt user', user)
+            // console.log('=>>> jwt token from api auth route page', token)
+            // console.log('=>>> jwt user from api auth route page', user)
 
             // Initial sign in
             if (user) {
@@ -127,8 +124,8 @@ export const authOptions = {
 
             // Return previous token if the access token has not expired yet
             // if (Date.now() < token.expiresAt) {
-            // console.log('=>>> token expired Date.now()', Date.now())
-            // console.log('=>>> token expired token.expiresAt', token.expiresAt)
+            // console.log('=>>> token expired Date.now()', Date.now() from api auth route page)
+            // console.log('=>>> token expired token.expiresAt from api auth route page', token.expiresAt)
             //   return token;
             // }
 
@@ -139,10 +136,10 @@ export const authOptions = {
         },
 
         async session({ session, token }) {
-            console.log("=>>> session...");
+            console.log("=>>> session from api auth route page...");
 
-            // console.log('=>>> session session', session)
-            // console.log('=>>> session token', token)
+            // console.log('=>>> session session from api auth route page', session)
+            // console.log('=>>> session token from api auth route page', token)
 
             if (token) {
                 session.accessToken = token.accessToken;
@@ -151,6 +148,9 @@ export const authOptions = {
             }
             return session;
         },
+    },
+    pages: {
+        newUser: "/new-user",
     },
     debug: true,
 };
