+++ app/components/Header.jsx
@@ -57,9 +57,9 @@ function Header() {
         storeUserAgent();
     }, []);
 
-    // console.log('=>>> get login status', status);
-    console.log('=>>> get login session', session);
-    // console.log('=>>> get login token', session?.accessToken);
+    console.log('=>>> get login status from header page', status);
+    console.log('=>>> get login session from header page', session);
+    // console.log('=>>> get login token from header page', session?.accessToken);
 
     useEffect(() => {
         storeUserAgent();
